Change of semantics of `ShardPositionsUpdate`
It now only contains the shards that are received an update, and the
event itself is only emitted if at least a shard received an update.

Closes #4500

+++ quickwit/quickwit-ingest/src/ingest_v2/state.rs
@@ -185,6 +185,11 @@ impl FullyLockedIngesterState<'_> {
     /// Deletes the shard identified by `queue_id` from the ingester state. It removes the
     /// mrecordlog queue first and then removes the associated in-memory shard and rate trackers.
     pub async fn delete_shard(&mut self, queue_id: &QueueId) {
+        // This if-statement is here to avoid needless log.
+        if self.inner.shards.contains_key(queue_id) {
+            // No need to do anything. This queue is not on this ingester.
+            return;
+        }
         match self.mrecordlog.delete_queue(queue_id).await {
             Ok(_) | Err(DeleteQueueError::MissingQueue(_)) => {
                 self.shards.remove(queue_id);
